+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Buffomat Classic 2022.9.5
 
 - [Shaman] Water Walking, Water Breath now in Buffomat, require no reagent with Wotlk Glyphs
+- [Shaman] Spellhance weapon enchants option (on by default) to enchant offhand with 1 rank lower flametongue.
+- [Priest] Vampiric Embrace is now a buff.
 
 ## Buffomat Classic 2022.9.4, 4.1
 - [Death Knight] Horn of Winter; Bone Shield